# Conflicts:
#	drivers/common/opentimers.c
#	openapps/cjoin/cjoin.c
#	openstack/02a-MAClow/IEEE802154E.c
#	openstack/02b-MAChigh/schedule.c
#	openstack/03b-IPv6/forwarding.c
#	projects/common/01bsp_radio_rx/01bsp_radio_rx.c
#	projects/common/02drv_openserial/02drv_openserial.c
#	projects/openmote-cc2538/03oos_openwsn/03oos_openwsn.ewp
#	projects/telosb/03oos_macpong/03oos_macpong.ewp

+++ bsp/boards/iot-lab_M3/spi.c
@@ -105,10 +105,10 @@ void spi_setCallback(spi_cbt cb) {
 #endif
 
 void spi_txrx(uint8_t*     bufTx,
-              uint8_t      lenbufTx,
+              uint16_t     lenbufTx,
               spi_return_t returnType,
               uint8_t*     bufRx,
-              uint8_t      maxLenBufRx,
+              uint16_t     maxLenBufRx,
               spi_first_t  isFirst,
               spi_last_t   isLast) {
